+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ### [Report an Issue](https://github.com/atlassian/atlascode/issues)
 ---
 
+## What's new in 3.7.0
+
+### Improvements
+- Badge notifications may be displayed if user is in an unauthenticated state. 
+
 ## What's new in 3.6.6
 
 - Internal changes only